refactor: move model trpg api to shared
