+++ src/portal/routes/trpg/recruit/list/index.tsx
@@ -2,7 +2,7 @@ import React, { useState, useCallback, useEffect } from 'react';
 import { TMemo } from '@shared/components/TMemo';
 import { RecruitItem } from '../RecruitItem';
 import MasonryLayout from 'react-masonry-layout';
-import { RecruitItemType, fetchAllRecruitList } from '@portal/model/trpg';
+import { RecruitItemType, fetchAllRecruitList } from '@shared/model/trpg';
 import { Layout, Empty, Button } from 'antd';
 import Loading from '@portal/components/Loading';
 import { RecruitCreateBtn } from './CreateBtn';
